Move MathSciNetId to identifier
