+++ src/main/java/org/jabref/model/entry/identifier/MathSciNetId.java
@@ -1,4 +1,4 @@
-package org.jabref.model.entry;
+package org.jabref.model.entry.identifier;
 
 import java.util.Objects;
